External lib #includes use <>, local #includes "" Fix some sprintf format warnings for 32/64 (%d with size_t) Don't specify scintilla lib in CMake when using wx-config.

+++ plotctrl/src/plotctrl.cpp
@@ -1066,7 +1066,7 @@ void wxPlotCtrl::CreateKeyString()
         else if (GetFunctionCurve(n) != NULL)
             key = GetFunctionCurve(n)->GetFunctionString();
         else
-            key.Printf(wxT("Curve %d"), n);
+            key.Printf(wxT("Curve %d"), (int)n);
 
         m_keyString += (key + wxT("\n"));
     }
